+++ ark/type/__tests__/traverse.test.ts
@@ -129,16 +129,14 @@ contextualize(() => {
 
 	it("multi", () => {
 		const naturalNumber = type("integer>0")
-		attest(naturalNumber(-1.2).toString()).snap(
-			`-1.2 must be...
+		attest(naturalNumber(-1.2).toString()).snap(`(-1.2) must be...
   • an integer
-  • more than 0`
-		)
+  • more than 0`)
 		const naturalAtPath = type({
 			natural: naturalNumber
 		})
 		attest(naturalAtPath({ natural: -0.1 }).toString()).snap(
-			`natural -0.1 must be...
+			`natural (-0.1) must be...
   • an integer
   • more than 0`
 		)
@@ -170,7 +168,7 @@ isAdmin must be false, null or true (was 1)`)
 		}).narrow(
 			(d, ctx) =>
 				d.password === d.repeatPassword ||
-				ctx.invalid({
+				ctx.reject({
 					expected: "identical to password",
 					actual: null,
 					relativePath: ["repeatPassword"]